# api/openapi.yaml
openapi: 3.0.3
info:
  title: Monelog API
  version: 1.0.0
  description: HTTP API for Monelog authentication and account profile management.
servers:
  - url: /
security:
  - BearerAuth: []
paths:
...
  /api/v1/auth/refresh:
    post:
      operationId: refresh
      summary: Rotate a refresh session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Rotated session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/AuthenticationFailed'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /api/v1/me:
    get:
      operationId: getMe
      summary: Get the current profile
      responses:
        '200':
          description: Current user profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/AuthenticationFailed'
    patch:
      operationId: updateMe
      summary: Update the current profile timezone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileRequest'
      responses:
        '200':
          description: Updated user profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/AuthenticationFailed'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      operationId: deleteMe
      summary: Soft-delete the current account
      parameters:
        - name: If-Match
          in: header
          required: true
          schema:
            type: string
          description: Current profile version, for example `"3"`.
      responses:
        '204':
          description: Account soft-deleted.
        '401':
          $ref: '#/components/responses/AuthenticationFailed'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RefreshRequest:
      type: object
      additionalProperties: false
      required: [client_type]
      properties:
        client_type:
          type: string
          enum: [native, web]
        refresh_token:
          type: string
          description: Required only for native clients; browser clients use an HttpOnly cookie.
    ProfileRequest:
      type: object
      additionalProperties: false
      required: [timezone, version]
      properties:
        timezone:
          type: string
          example: Asia/Jakarta
        version:
          type: integer
          format: int32
          minimum: 1
    User:
      type: object
      required: [id, email, role, timezone, currency, isDelete, version]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [user, admin]
        timezone:
          type: string
        currency:
          type: string
          enum: [IDR]
        isDelete:
          type: boolean
        version:
          type: integer
          format: int32
    UserResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/User'
    LoginResponse:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
          description: Returned only for native clients.
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          format: int64
          example: 86400
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Error'
  responses:
    AuthenticationFailed:
      description: Authentication failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Request conflicts with current state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Request is not permitted.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'